requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.2
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
docstring_parser==0.17.0
exceptiongroup==1.3.0
fastapi==0.119.1
filetype==1.2.0
frozenlist==1.8.0
google==3.0.0
google-ai-generativelanguage==0.9.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-aiplatform==1.121.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.14.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-cloud-translate==3.22.0
google-crc32c==1.7.1
google-genai==1.45.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.1
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pymilvus==2.6.2
PyMuPDF==1.26.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
gunicorn
zstandard==0.25.0
redis
reportlab
python-docx
